Fusion Dev =>Trunk
git-tfs-id: [http://172.17.236.41:8080/]$/AdExpress International Demo/Trunk;C25541

+++ DLL/TNS.AdExpress.Web.Controls/Selections/RecapAdvertiserSelectionWebControl.cs
@@ -142,7 +142,7 @@ namespace TNS.AdExpress.Web.Controls.Selections
 
 				//Construction de la liste de checkbox
 				foreach(DataRow currentRow in dsListAdvertiser.Tables[0].Rows) {
-					this.Items.Add(new System.Web.UI.WebControls.ListItem(currentRow["advertiser"].ToString(),((Int64)currentRow["id_advertiser"]).ToString()));
+					this.Items.Add(new System.Web.UI.WebControls.ListItem(currentRow["advertiser"].ToString(),currentRow["id_advertiser"].ToString()));
 				}
 			//}
 		}
